// collect/company/new-cloud/cloud-portal/src/pages/cloud-home/ch-change-energy-situation/service/ch-change-energy-situation.api.ts
/**
 * 碳排放强度echarts
 * @param projectCode 节能项目编码
 * @param value 实施项目
 * @param projectName 节能项目名称
 * @param savingAmount 已节约金额
 */
export interface TResChangeEnergySituationObj {
  projectCode: string;
  projectCount: string;
  projectName: string;
  savingAmount: string;
}